add utils for extracting extensions
